ImpedanceController: fix serious bug in position error calculation.
Underlying the known issue with the wrong stiffness sign was a more
serious bug. The inversion of both measured and desired position
resulted in the 'diff' operation to be carried out with arguments
represented in incompatible reference frames. The problem wasn't
visible for small errors.

Thanks to Bert Willaert for spotting this bug and Tinne De Laet for
assisting in solving it.